% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_stat_unweight.R
\name{kappa_stat_unweight}
\alias{kappa_stat_unweight}
\title{Computes an unweighted kappa statistic}
\usage{
kappa_stat_unweight(table, ci = FALSE, alpha = 0.05)
}
\arguments{
\item{table}{A square confusion matrix}

\item{ci}{Return a confidence interval}

\item{alpha}{Confidence level for condifence interval}
}
\description{
Computes an unweighted kappa statistic
}
\examples{
dgam_plot(a_square_matrix)
}
